PVS scan audio module

+++ modules/audio/builtin/cpp/audiowriteBuiltin.cpp
@@ -75,7 +75,7 @@ Nelson::AudioGateway::audiowriteBuiltin(Evaluator* eval, int nLhs, const ArrayOf
         }
         if (!validFieldname) {
             wchar_t buffer[4096];
-            swprintf(buffer, 4096, std::wstring(ERROR_WRONG_ARGUMENT_X_VALUE).c_str(), i);
+            swprintf(buffer, 4096, std::wstring(ERROR_WRONG_ARGUMENT_X_VALUE).c_str(), (int)i);
             Error(std::wstring(buffer));
         }
     }
